backend/requirements.txt
Flask
Flask-Cors
mysql-connector-python
pandas
scikit-learn